// src/types/ebay.ts
export interface EbayConfig {
  clientId: string;
  clientSecret: string;
  redirectUri?: string;
  locale?: LocaleEnum;
  marketplaceId?: string;
  contentLanguage?: string;
  environment: 'production' | 'sandbox';
  accessToken?: string;
  refreshToken?: string;
  appAccessToken?: string;
}